Make equatable conformance public
