+++ Sources/JWTKit/ECDSA/ECDSAError.swift
@@ -4,5 +4,4 @@ enum ECDSAError: Error {
     case signFailure
     case noPublicKey
     case noPrivateKey
-    case wrongCurve
 }
