+++ Sources/JWTKit/RSA/RSASigner.swift
@@ -36,10 +36,7 @@ struct RSASigner: JWTAlgorithm, CryptoSigner {
         let digest = try self.digest(plaintext)
         let signature = _RSA.Signing.RSASignature(rawRepresentation: signature)
 
-        guard let publicKey = key.privateKey?.publicKey ?? key.publicKey else {
-            throw JWTError.signingAlgorithmFailure(RSAError.publicKeyRequired)
-        }
-        return publicKey.isValidSignature(signature, for: digest, padding: padding)
+        return key.publicKey.isValidSignature(signature, for: digest, padding: padding)
     }
 }
